requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
colorama==0.4.6
cron-descriptor==1.4.5
decorator==5.1.1
dj-database-url==2.3.0
Django==4.2.16
django-celery-beat==2.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-filter==24.2
django-leaflet==0.31.0
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.2
frozenlist==1.5.0
future==1.0.0
geocoder==1.38.1
geoip2==4.8.1
gunicorn==22.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
maxminddb==2.6.2
GDAL==3.6.2
numpy>=1.26.0

mkl-service
# mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
# mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
multidict==6.1.0
# numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp39-cp39-win_amd64.whl#sha256=f284633067fde4bb078a09085a1c9e99a3218c129f5312b7b3d76c9340442a25
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
ratelim==0.1.6
redis==5.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
yarl==1.17.2
